quick commit

+++ HI_FAST/test_light.py
@@ -17,7 +17,7 @@ import timeit
 from ctypes import CDLL, POINTER, c_int, c_float, c_double
 #import test as mymodule
 mymodule = CDLL('./test.so')
-mymodule.make_sphere.argtypes = [ POINTER(c_int), POINTER(c_float), POINTER(c_float),POINTER(c_float)]
+mymodule.make_sphere.argtypes = [POINTER(c_int), POINTER(c_float), POINTER(c_float),POINTER(c_float)]
 
 from timeit import default_timer as timer
 rank = "0"
@@ -102,8 +102,8 @@ def plot_coldgas(z):
         if not index in gal:
             (nTrees[index],nGals[index],nTreeGals[index],gal[index]) = read_lgal.readsnap_lgal_advance(model_paths[i],file_prefix,0,8,filter[i],dt[i],1)
 
-        pos_sphere = numpy.empty((nGals[index]*8,3),dtype=numpy.float32)
-        pos =  gal[index]['Pos']
+        pos_sphere = numpy.empty((nGals[index]*8,3),dtype=c_float)
+        pos =  gal[index]['Pos'].astype(c_float)
         print pos
         
         mymodule.make_sphere(c_int(nGals[index]),c_float(500.0),pos.ctypes.data_as(POINTER(c_float)),pos_sphere.ctypes.data_as(POINTER(c_float)))
